+++ lib/domain/model/id.dart
@@ -195,13 +195,21 @@ class Id implements IdApi {
            if (a.type.base == 'String') {
              String attribute = _attributeMap[a.code];
              compare = attribute.compareTo(id.getAttribute(a.code));
-           } else if (a.type.base == 'DateTime') {
-             DateTime attribute = _attributeMap[a.code];
-             compare = attribute.compareTo(id.getAttribute(a.code));
            } else if (a.type.base == 'num' ||
              a.type.code == 'int' || a.type.code == 'double') {
              num attribute = _attributeMap[a.code];
              compare = attribute.compareTo(id.getAttribute(a.code));
+           } else if (a.type.base == 'DateTime') {
+             DateTime attribute = _attributeMap[a.code];
+             compare = attribute.compareTo(id.getAttribute(a.code));
+           } else if (a.type.base == 'Uri') {
+             Uri attribute = _attributeMap[a.code];
+             compare = attribute.toString().compareTo(
+                 id.getAttribute(a.code).toString());
+           } else if (a.type.base == 'bool') {
+             bool attribute = _attributeMap[a.code];
+             compare = attribute.toString().compareTo(
+                 id.getAttribute(a.code).toString());
            } else {
              String msg = 'cannot compare then order on this type';
              if (a.concept != null) {